cmd: Use HTTP roundtripper from scorecard
When scorecard moved to v2, support for older GitHub clients in
`CheckRequest` was dropped, which means any means of GitHub
authentication instantiated within allstar gets dropped as well.

Specifically, this affects the binary artifacts policy check.

Here, we opt to use scorecard's HTTP roundtripper in
`NewGHClients`, which should allow us to seed GitHub credentials
in the same way that scorecard does.

Unfortunately, this has the mild side-effect of temporarily
introducing zap as a logger for a small portion of the code (but
this will be removed again once we move to scorecard/v4).

Signed-off-by: Stephen Augustus <foo@auggie.dev>