Overhaul of the GitHub Actions.
Grouping based on triggers, so: PR, postmerge (push), and release (tag). PR and
postmerge are read-only permissions, so they just use major version tags for
actions to avoid dependabot annyoance. The release workflow uses hash
pinning. I've removed the PAT secret for the scorecard action, so it will just
be using read-only access. I've updated to Go 1.19, but not the deploy scripts
yet.