Change cache to avoid memory use
Orignally, the cache was intended to be long lived to handle incoming webhooks
at any time. Currently, we are just polling, and just need the cache to handle
a single "EnforceAll" run, where we hit the same paths multiple times in that
run. Therefore, change the cache to be per-installation, and free it after each
"EnforceAll".

Signed-off-by: Jeff Mendoza <jlm@jlm.name>